% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_boolean_args}
\alias{check_boolean_args}
\title{Check Boolean Arguments}
\usage{
check_boolean_args(cyclized = NULL, clipped = NULL, glycosylation = NULL)
}
\arguments{
\item{cyclized}{A logical value indicating whether cyclization is enabled}

\item{clipped}{A logical value indicating whether clipping is enabled}

\item{glycosylation}{A logical value indicating whether glycosylation is enabled}
}
\value{
Nothing. Called for side effects.
}
\description{
Validates that specified arguments (Cyclization, clipping, and glycosylation)
are boolean values (TRUE/FALSE). Provides specific error messages for any
non-boolean arguments.
}
\keyword{internal}
