% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_chem_mod_args}
\alias{check_chem_mod_args}
\title{Validate Chemical Modification Arguments}
\usage{
check_chem_mod_args(hc_chem_mod = NA, lc_chem_mod = NA)
}
\arguments{
\item{hc_chem_mod}{Character string or NA. Chemical modification for heavy chain.}

\item{lc_chem_mod}{Character string or NA. Chemical modification for light chain.}
}
\value{
None. Raises an error if validation fails.
}
\description{
Internal function to validate the format of chemical modification strings for
heavy and light chains.
}
\keyword{internal}
