% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input}
\alias{check_input}
\title{Standard one-letter amino acid codes}
\usage{
check_input(one_letter_input)
}
\arguments{
\item{one_letter_input}{A character string containing the amino acid sequence}
}
\value{
NULL invisibly. Raises an error if validation fails
}
\description{
Standard one-letter amino acid codes
}
\keyword{(HC)}
\keyword{(LC)}
\keyword{HC/LC}
\keyword{Internal}
\keyword{Validates}
\keyword{acid}
\keyword{amino}
\keyword{and}
\keyword{antibody}
\keyword{chain}
\keyword{codes.}
\keyword{contains}
\keyword{correct}
\keyword{for}
\keyword{format}
\keyword{function}
\keyword{heavy}
\keyword{in}
\keyword{input}
\keyword{internal}
\keyword{is}
\keyword{light}
\keyword{one-letter}
\keyword{only}
\keyword{or}
\keyword{sequence}
\keyword{the}
\keyword{to}
\keyword{valid}
\keyword{validate}
\keyword{whether}
