% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{XXX}
\usage{
get_params(hc_seq, hc_cyclized, hc_clipped)
}
\arguments{
\item{hc_seq}{A string with hc amino acids}

\item{hc_cyclized}{Boolean}

\item{hc_clipped}{Boolean}
}
\value{
vector
}
\description{
XXX
}
