% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params_monoclonal}
\alias{get_params_monoclonal}
\title{XXX}
\usage{
get_params_monoclonal(hc_seq, lc_seq)
}
\arguments{
\item{hc_seq}{A string with hc amino acids}

\item{lc_seq}{A string with lc amino acids}
}
\value{
vector
}
\description{
XXX
}
