% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_seq_element_comp}
\alias{get_seq_element_comp}
\title{Get elemental composition for specified molecules}
\usage{
get_seq_element_comp(seq)
}
\arguments{
\item{seq}{Character vector of molecule names to look up
\itemize{
\item molecule: Character. Name of the molecule
\item carbon: Integer. Number of carbon atoms
\item hydrogen: Integer. Number of hydrogen atoms
\item nitrogen: Integer. Number of nitrogen atoms
\item oxygen: Integer. Number of oxygen atoms
\item sulphur: Integer. Number of sulfur atoms
}}
}
\value{
A data frame containing rows from element_composition matching the
input molecules
}
\description{
Internal function that retrieves the elemental composition data for a given
set of molecules from the element_composition dataset.
}
\examples{
\dontrun{
# Get composition for multiple molecules
get_seq_element_comp(c("ala", "gly", "ser"))

# Get composition for a single molecule
get_seq_element_comp("water")
}
}
\keyword{internal}
