% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_clearance.R
\name{mab_clearance}
\alias{mab_clearance}
\title{Predicts clearance of human IgG1 based on variable fragments and CDRs of
heavy and light chains}
\usage{
mab_clearance(hc_fv, lc_fv, cdr_h3, cdr_l1, cdr_l3)
}
\arguments{
\item{hc_fv}{A string containing heavy variable fragment amino acids (one letter code)}

\item{lc_fv}{A string containing light variable fragment amino acids (one letter code)}

\item{cdr_h3}{A string containing heavy CDR3 amino acids (one letter code)}

\item{cdr_l1}{A string containing heavy CDL1 amino acids (one letter code)}

\item{cdr_l3}{A string containing heavy CDL1 amino acids (one letter code)}
}
\value{
A tibble with clearance parameters
}
\description{
Predicts clearance of human IgG1 based on variable fragments and CDRs of
heavy and light chains
}
\examples{
h_fv <- "EVQLVESGGGLVQPGGSLRLSCAASGFNIKDTYIHWVRQAPGKGLEWVARIYPTNGYTR"
l_fv <- "DIQMTQSPSSLSASVGDRVTITCRASQDVNTAVAWYQQKPGKAPKLLIYSAS"
cdr_h3 = "WGGDGFYAMDY"
cdr_l1 = "RASQDVNTAVA"
cdr_l3 = "SASFLYS"
clearance_df <- mab_clearance(h_fv, l_fv, cdr_h3, cdr_l1, cdr_l3)
clearance_df
}
