% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_comp.R
\name{mab_comp}
\alias{mab_comp}
\title{mab_comp returns a tibble with the amino acid compositions of intact
monoclonal antibody (cys residues are reduced). Charges on each amino
acid is assigned based on pI of the antibody}
\usage{
mab_comp(hc_seq, lc_seq)
}
\arguments{
\item{hc_seq}{A string containing heavy chain amino acids (one letter code)}

\item{lc_seq}{A string containing light chain amino acids (one letter code)}
}
\value{
A tibble with amino acid compositions
}
\description{
mab_comp returns a tibble with the amino acid compositions of intact
monoclonal antibody (cys residues are reduced). Charges on each amino
acid is assigned based on pI of the antibody
}
\examples{
heavy <- "EVQLVESGGGLVQPGGSLRLSCAASGFNIKDTYIHWVRQAPGKGLEWVARIYPTNGYTRYA"
light <- "DIQMTQSPSSLSASVGDRVTITCRASQDVNTAVAWYQQKPGKAPKLLIYSASFLYSGVPSR"
comp_tbl <- mab_comp(heavy, light)
head(comp_tbl)
}
