% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_viscosity.R
\name{mab_viscosity}
\alias{mab_viscosity}
\title{Calculate charge and viscosity (in cP) of human IgG1 antibody at a pH of 5.5
Equation: η, cP(180mg/mL, 25°C) = 10^(0.15 + 1.26(0.60)∗ϕ − 0.043(0.047)∗q − 0.020(0.015)∗qsym)
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4284567/#eq1}
\usage{
mab_viscosity(hc_fv, lc_fv)
}
\arguments{
\item{hc_fv}{A string containing variable heavy chain amino acids (one letter code)}

\item{lc_fv}{A string containing variable light chain amino acids (one letter code)}
}
\value{
A numeric value of predicted viscosity in cP
}
\description{
Calculate charge and viscosity (in cP) of human IgG1 antibody at a pH of 5.5
Equation: η, cP(180mg/mL, 25°C) = 10^(0.15 + 1.26(0.60)∗ϕ − 0.043(0.047)∗q − 0.020(0.015)∗qsym)
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4284567/#eq1
}
\examples{
h_fv <- "EVQLVESGGGLVQPGGSLRLSCAASGFNIKDTYIHWVRQAPGKGLEWVARIYPTNGYTR"
l_fv <- "DIQMTQSPSSLSASVGDRVTITCRASQDVNTAVAWYQQKPGKAPKLLIYSAS"
visc <- mab_viscosity(h_fv, l_fv)
visc
}
