% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabParams-package.R
\docType{package}
\name{mabParams-package}
\alias{mabParams}
\alias{mabParams-package}
\title{mabParams: Calculate Mass for Antibody based Therapeutic Applications}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shubhamdutta26/mabParams}
  \item Report bugs at \url{https://github.com/shubhamdutta26/mabParams/issues}
}

}
\author{
\strong{Maintainer}: Shubham Dutta \email{shubhamdutta26@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [copyright holder]

}
\keyword{internal}
