% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chem_formula.R
\name{parse_chem_formula}
\alias{parse_chem_formula}
\title{Converts chemical formula into a tibble with names and element counts. Only
works for crtain elements. Add more elements if needed in the named vector
inside the function}
\usage{
parse_chem_formula(formula)
}
\arguments{
\item{formula}{A character vector og length 1 (e.g. C2H5OH for ethanol)}
}
\value{
A tibble/ dataframe
}
\description{
Converts chemical formula into a tibble with names and element counts. Only
works for crtain elements. Add more elements if needed in the named vector
inside the function
}
