% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_params.R
\name{mab_params}
\alias{mab_params}
\title{Calculate mass parameters of monoclonal antibodies}
\usage{
mab_params(
  hc_seq,
  lc_seq,
  hc_cyclized = FALSE,
  hc_clipped = FALSE,
  lc_glycosylation = FALSE,
  hc_chem_mod = NA,
  lc_chem_mod = NA,
  n_disulphides = 16L,
  n_hc_disulphides = 4L,
  n_lc_disulphides = 2L,
  glycans = NA,
  mab = "mab1"
)
}
\arguments{
\item{hc_seq}{A string containing heavy chain amino acids (one letter code)}

\item{lc_seq}{A string containing light chain amino acids (one letter code)}

\item{hc_cyclized}{A boolean value for heavy chain cyclization}

\item{hc_clipped}{A boolean value for heavy chain clipping}

\item{lc_glycosylation}{A boolean value for light chain glycosylation}

\item{hc_chem_mod}{A character vector of chemical formula of length 1}

\item{lc_chem_mod}{A character vector of chemical formula of length 1}

\item{n_disulphides}{Number of total disulphides}

\item{n_hc_disulphides}{Number of total heavy chain disulphides}

\item{n_lc_disulphides}{Number of total light chain disulphides}

\item{glycans}{A csv file that contains glycan parameters}

\item{mab}{Name of the antibody; default is "mab1"}
}
\value{
A dataframe/ tibble with antibody mass parameters
}
\description{
Calculate mass parameters of monoclonal antibodies
}
\examples{
heavy <- "EVQLVESGGGLVQPGGSLRLSCAASGFNIKDTYIHWVRQAPGKGLEWVARIYPTNGYTRYA"
light <- "DIQMTQSPSSLSASVGDRVTITCRASQDVNTAVAWYQQKPGKAPKLLIYSASFLYSGVPSR"
glycans_path <- system.file("extdata", "glycans.csv", package = "mabParams")
params_tbl <- mab_params(heavy, light, glycans = glycans_path)
head(params_tbl)
}
