% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_params.R
\name{mab_params}
\alias{mab_params}
\title{Calculate mass parameters of monoclonal antibodies}
\usage{
mab_params(
  hc_seq,
  lc_seq,
  hc_cyclized = F,
  hc_clipped = F,
  lc_glycosylation = F,
  hc_chem_mod = NA,
  lc_chem_mod = NA,
  n_disulphides = 16L,
  n_hc_disulphides = 4L,
  n_lc_disulphides = 2L,
  mab = "mab1"
)
}
\arguments{
\item{hc_seq}{A string with hc amino acids}

\item{lc_seq}{A string with lc amino acids}

\item{hc_cyclized}{A boolean}

\item{hc_clipped}{A boolean}

\item{lc_glycosylation}{A boolean}

\item{hc_chem_mod}{A string}

\item{lc_chem_mod}{A string}

\item{n_disulphides}{Number of total disulphides}

\item{n_hc_disulphides}{Number of total HC disulphides}

\item{n_lc_disulphides}{Number of total LC disulphides}

\item{mab}{Name of mab; default is mab1}
}
\value{
vector
}
\description{
Calculate mass parameters of monoclonal antibodies
}
