% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_params.R
\name{mab_params}
\alias{mab_params}
\title{XXX}
\usage{
mab_params(
  hc_seq,
  lc_seq,
  hc_cyclized,
  hc_clipped,
  n_disulphides = 16L,
  n_hc_disulphides = 4L,
  n_lc_disulphides = 2L,
  mab = "mab1"
)
}
\arguments{
\item{hc_seq}{A string with hc amino acids}

\item{lc_seq}{A string with lc amino acids}

\item{hc_cyclized}{A boolean}

\item{hc_clipped}{A boolean}

\item{n_disulphides}{Number of total disulphides}

\item{n_hc_disulphides}{Number of total HC disulphides}

\item{n_lc_disulphides}{Number of total LC disulphides}

\item{mab}{Name of mab; default is mab1}
}
\value{
vector
}
\description{
XXX
}
