% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabParams-package.R
\docType{package}
\name{mabParams-package}
\alias{mabParams}
\alias{mabParams-package}
\title{mabParams: Calculate Mass for Antibody based Therapeutic Applications}
\description{
A mass calculator that facilitates the straightforward determination of average or monoisotopic masses and elemental compositions for therapeutic glycoproteins, such as monoclonal antibodies (mAb), bispecific antibodies (bsAb), and antibody-drug conjugates (ADC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shubhamdutta26/mabParams}
  \item Report bugs at \url{https://github.com/shubhamdutta26/mabParams/issues}
}

}
\author{
\strong{Maintainer}: Shubham Dutta \email{shubhamdutta26@gmail.com} (\href{https://orcid.org/0000-0001-8484-0717}{ORCID}) [copyright holder]

}
\keyword{internal}
